init
This is just yet another PKGBREAK generator, providing better control
over the package selecting process to user.

This supports a PKGREBUILD syntax, for example:
PKGREBUILD__ABI="sodep revdep \
	path::usr/lib/python.*/.* \
	path::.*\\.pyc \
	section::lang-python"
Will include:
- All shared-library dependents.
- All PKGDEP dependents.
- All packages providing files under /usr/lib/pythonX.X
- All packages under lang-python section.
The third is important for upgrading Python versions.

This also allows user to choose to bump REL of all broken packages but
not committing, in order to avoid flashing the blinking firmware to their
GPG keys.

Another improvement is the ability to parse APML files more robustly.
For example, breaker often generates PKGBREAK relationships like
"XXX<=1:"${_UPSTREAM_VER}"", while BreakIt will print the evaluated
version string in these cases.
The two existing APML parsing libraries are both written in Rust and it
seems to be a bit hard to invoke them in Python.

Another feature is being capable to write the generated PKGBREAK back
into defines files. With the power of LST, this will not break the
existing code style of defines files.

However, this is not an alternative to breaker, for that not all
features in breaker are implemented here. I am not sure about what I
missed but there must be something I missed. So use this tool at your
own risk.